first- commit : Custom Exception

+++ src/main/java/com/stackroute/muzix/service/MusicServiceImpl.java
@@ -1,10 +1,13 @@
 package com.stackroute.muzix.service;
 
 import com.stackroute.muzix.domain.Music;
+import com.stackroute.muzix.exceptions.TrackAlreadyExistsException;
+import com.stackroute.muzix.exceptions.TrackNotFoundException;
 import com.stackroute.muzix.repository.MusicRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,7 +27,11 @@ public class MusicServiceImpl implements MusicService {
     }
 
     @Override
-    public Music updateUser(Music user) {
+    public Music updateUser(Music user) throws TrackAlreadyExistsException {
+        if(musicRepository.existsById(user.getTrackId()))
+        {
+            throw new TrackAlreadyExistsException("Track Already exists with Id "+ user.getTrackId());
+        }
         Optional<Music> user1 = musicRepository.findById(user.getTrackId());
 
         Music user2 = user1.get();
@@ -35,16 +42,28 @@ public class MusicServiceImpl implements MusicService {
     }
 
     @Override
-    public void deleteUser(int trackId) {
+    public void deleteUser(int trackId) throws TrackNotFoundException {
+        if(!musicRepository.existsById(trackId))
+        {
+            throw new TrackNotFoundException("Track Not Found with Id "+ trackId);
+        }
         Optional<Music> user1 = musicRepository.findById(trackId);
         Music user2 = user1.get();
         musicRepository.delete(user2);
     }
 
     @Override
-    public Music findById(int trackId) {
+    public Music findById(int trackId) throws TrackNotFoundException
+    {
+
+        if(!musicRepository.existsById(trackId))
+        {
+            throw new TrackNotFoundException("Track Not Found with Id "+ trackId);
+        }
         Optional<Music> userId = musicRepository.findById(trackId);
+        System.out.println(userId.toString());
         Music user2 = userId.get();
+
         return user2;
     }
 
@@ -54,8 +73,12 @@ public class MusicServiceImpl implements MusicService {
     }
 
     @Override
-    public List<Music> queryString(String name) {
+    public List<Music> queryString(String name) throws TrackNotFoundException {
         List<Music> list = musicRepository.getMusicInfoByName(name);
+        if(list.isEmpty())
+        {
+            throw new TrackNotFoundException("No tracks found with name " + name);
+        }
         return list;
     }
